rename: title -> memberTitle 패키지명 수정
