+++ src/main/java/com/idea5/four_cut_photos_map/domain/member/entity/Member.java
@@ -1,8 +1,8 @@
 package com.idea5.four_cut_photos_map.domain.member.entity;
 
 import com.idea5.four_cut_photos_map.domain.like.entity.Like;
-import com.idea5.four_cut_photos_map.domain.title.entity.MemberTitle;
-import com.idea5.four_cut_photos_map.domain.title.entity.MemberTitleLog;
+import com.idea5.four_cut_photos_map.domain.memberTitle.entity.MemberTitle;
+import com.idea5.four_cut_photos_map.domain.memberTitle.entity.MemberTitleLog;
 import com.idea5.four_cut_photos_map.global.base.entity.BaseEntity;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
