+++ src/main/java/com/idea5/four_cut_photos_map/domain/member/service/MemberService.java
@@ -5,9 +5,9 @@ import com.idea5.four_cut_photos_map.domain.member.dto.response.MemberInfoResp;
 import com.idea5.four_cut_photos_map.domain.member.dto.response.MemberWithdrawlResp;
 import com.idea5.four_cut_photos_map.domain.member.entity.Member;
 import com.idea5.four_cut_photos_map.domain.member.repository.MemberRepository;
-import com.idea5.four_cut_photos_map.domain.title.entity.MemberTitle;
-import com.idea5.four_cut_photos_map.domain.title.service.TitleService;
-import com.idea5.four_cut_photos_map.domain.title.entity.MemberTitleLog;
+import com.idea5.four_cut_photos_map.domain.memberTitle.entity.MemberTitle;
+import com.idea5.four_cut_photos_map.domain.memberTitle.service.TitleService;
+import com.idea5.four_cut_photos_map.domain.memberTitle.entity.MemberTitleLog;
 import com.idea5.four_cut_photos_map.global.common.RedisDao;
 import com.idea5.four_cut_photos_map.security.jwt.JwtProvider;
 import lombok.RequiredArgsConstructor;
