+++ src/main/java/com/idea5/four_cut_photos_map/domain/memberTitle/entity/TitleType.java
@@ -1,4 +1,4 @@
-package com.idea5.four_cut_photos_map.domain.title.entity;
+package com.idea5.four_cut_photos_map.domain.memberTitle.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
