+++ src/main/java/com/idea5/four_cut_photos_map/domain/memberTitle/repository/TitleRepository.java
@@ -1,0 +1,7 @@
+package com.idea5.four_cut_photos_map.domain.memberTitle.repository;
+
+import com.idea5.four_cut_photos_map.domain.memberTitle.entity.MemberTitle;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TitleRepository extends JpaRepository<MemberTitle, Long> {
+}
